# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book-Store",
        "description": "Sistema Gerencial de Livros",
        "contact": {
            "name": "Josimar Domingos",
            "email": "josimardomingos@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/assuntos": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Listar todos os assuntos",
                "operationId": "b65bfc4a125a8df880e922c685bb247b",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Assunto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Criar um novo autor",
                "operationId": "2e7489bfd9849f0707489d0f09c8263a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/assuntos/{id}": {
            "get": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Mostrar detalhes de um autor específico",
                "operationId": "75184321a3ffcd363b75f3e8bd451468",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Atualizar os dados de um autor",
                "operationId": "3f9c17df8d02b0b4e11917bce2a048e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assunto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assunto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assuntos"
                ],
                "summary": "Excluir um autor",
                "operationId": "513e9acbc689a6e5f0855eb11ea43909",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful"
                    }
                }
            }
        },
        "/autores": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Listar todos os autores",
                "operationId": "eb27430fea956ab5ff5aaca0e1355868",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Autor"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Autores"
                ],
                "summary": "Criar um novo autor",
                "operationId": "79b1e52e74880ae8637d083d4705c17d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/autores/{id}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Mostrar detalhes de um autor específico",
                "operationId": "dd01080f500d7bbcb4d41eba034ebbff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Autores"
                ],
                "summary": "Atualizar os dados de um autor",
                "operationId": "9dc6b2621eb50caac0dc2bcd21888aa8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Autores"
                ],
                "summary": "Excluir um autor",
                "operationId": "fe487831afee2422c6afc92243a36d1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful"
                    }
                }
            }
        },
        "/livros": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Listar todos os livros",
                "operationId": "c2638f778b25c161b5b8189172216b67",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Livro"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Livros"
                ],
                "summary": "Criar um novo livro",
                "operationId": "052dc25bf67cc3710e476ca8c1f49397",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livro"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/livros/{id}": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Mostrar detalhes de um livro específico",
                "operationId": "5ee1c89a5977dcbe47813c762d44c15f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Livros"
                ],
                "summary": "Atualizar os dados de um livro",
                "operationId": "34f924b16d17d726283baad63b144290",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livro"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Livros"
                ],
                "summary": "Excluir um livro",
                "operationId": "20ff1e31d004d5d38f9e8c62a3568b1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Assunto": {
                "title": "Assunto",
                "properties": {
                    "codas": {
                        "type": "integer",
                        "example": "7"
                    },
                    "nome": {
                        "type": "string",
                        "example": "Assunto de teste"
                    }
                },
                "type": "object"
            },
            "Autor": {
                "title": "Autor",
                "properties": {
                    "codau": {
                        "type": "integer",
                        "example": "12"
                    },
                    "nome": {
                        "type": "string",
                        "example": "Autor de teste"
                    }
                },
                "type": "object"
            },
            "Livro": {
                "title": "Livro",
                "properties": {
                    "codl": {
                        "type": "integer",
                        "example": "17"
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Titulo de teste"
                    },
                    "editora": {
                        "type": "string",
                        "example": "Editora de teste"
                    },
                    "edicao": {
                        "type": "integer",
                        "example": "2"
                    },
                    "anopublicacao": {
                        "type": "integer",
                        "example": "2023"
                    }
                },
                "type": "object"
            }
        }
    }
}